langserve_api/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langchain>=0.0.352
langserve>=0.0.41
chromadb>=0.4.22
pydantic>=2.5.2
python-dotenv>=1.0.0
langgraph>=0.0.15 